fix npe in local mode and concurrent modification in AbstractNode

+++ engine/src/main/java/com/malhartech/lib/io/HttpOutputNode.java
@@ -60,8 +60,10 @@ public class HttpOutputNode extends AbstractNode
 
   @Override
   public void teardown() {
+    if (wsClient != null) {
+      wsClient.destroy();
+    }
     super.teardown();
-    wsClient.destroy();
   }
 
   @Override
